Move all source files into src/, remove old unused files, and misc clean ups
