+++ Makefile
@@ -6,4 +6,5 @@ test-browser:
 	node scripts/run-browser-tests-live-reload.js
 
 lint:
-	./node_modules/.bin/eslint --ignore-pattern tests/util/deepEqual.js tests/ addon/ storage/ store-engine.js util.js
+	./node_modules/.bin/eslint tests/* src/*      \
+		--ignore-pattern tests/util/deepEqual.js  \
