+++ Roadmap.md
@@ -41,33 +41,3 @@ Open pull requests & issues:
 - [ ] #138: `typeof window` check
 	- [ ] Reproduce: Create failing test case
 	- [ ] Fix failing test case
-
-Open issues
-
-
-v2.x.x:
--------
-
-Plugins system:
-- [ ] identify a few example plugins (mostly from issues)
-- [ ] Consider how https://github.com/evitolins/StoreFront.js does it
-- [ ] how are plugins included+loaded?
-- [ ] implement plugin system and a few example plugins
-- [ ] Features to consider
-	- [ ] Stored data structures (e.g #134, #136, #146)
-	- [ ] Encryption (#137)
-	- [ ] Backwards forEach (#144)
-	- [ ] #133: Handle namespace
-	- [ ] #126: localStorage vs sessionStorage
-	- [ ] #119: Fall back to in-memory storage
-	- [ ] #129: noop if no supported options found
-
-
-Examples to look at:
-- [ ] Consider using json3 instead of json2: https://bestiejs.github.io/json3/
-- [ ] https://github.com/evitolins/StoreFront.js
-- [ ] http://download.github.io/memorystorage/
-- [ ] https://github.com/nbubna/store
-- [ ] http://rhaboo.org/
-
-
