--- addon/all.js
@@ -1,10 +1,0 @@
-module.exports = {
-	'defaults': require('./defaults'),
-	'dump': require('./dump'),
-	'events': require('./events'),
-	'observe': require('./observe'),
-	'expire': require('./expire'),
-	'operations': require('./operations'),
-	'update': require('./update'),
-	'v1-backcompat': require('./v1-backcompat')
-}
