--- addon/all_tests.js
@@ -1,9 +1,0 @@
-module.exports = {
-	'defaults': require('./defaults_test'),
-	'dump': require('./dump_test'),
-	'events': require('./events_test'),
-	'expire': require('./expire_test'),
-	'update': require('./update_test'),
-	'operations': require('./operations_test'),
-	'v1-backcompat': require('./v1-backcompat_test')
-}
