+++ src/addon/all_tests.js
@@ -1,0 +1,10 @@
+module.exports = {
+	'defaults':      require('./defaults_test'),
+	'dump':          require('./dump_test'),
+	'events':        require('./events_test'),
+	'observe':       require('./observe_test'),
+	'expire':        require('./expire_test'),
+	'operations':    require('./operations_test'),
+	'update':        require('./update_test'),
+	'v1-backcompat': require('./v1-backcompat_test')
+}
