+++ src/addon/events_test.js
