+++ src/addon/observe_test.js
