+++ src/addon/operations_test.js
@@ -1,5 +1,5 @@
 var { each, map } = require('../util')
-var { deepEqual } = require('../tests/util')
+var { deepEqual } = require('../../tests/util')
 
 require('./operations')
