--- src/store+json2.js
@@ -1,2 +1,0 @@
-require('./json2')
-module.exports = require('./store')
